+++ dragonpoop_alpha/dragonpoop/gfx/model/model_loader/model_loader_ms3d/model_loader_ms3d_state_make_frames.h
@@ -27,7 +27,7 @@ namespace dragonpoop
         
         dpbuffer *b;
         model_ref *m;
-        std::vector<int> frame_times;
+        std::vector<float> frame_times;
         
     protected:
